refactor: firebase app provider, login styles
