+++ tournament_app/app/login/page.tsx
@@ -22,7 +22,7 @@ import * as z from "zod";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useAuthState } from "react-firebase-hooks/auth";
-import { auth } from "@/components/lib/firebase/clientApp";
+import { auth } from "@/components/lib/firebase/client-app";
 import { useRouter } from "next/navigation";
 
 const formSchema = z.object({
@@ -117,7 +117,7 @@ function page() {
               </Button>
             </form>
           </Form>
-          <div>
+          <div className="max-w-xs w-full flex flex-col items-center">
             <span className="flex w-full md:w-60 lg:w-80">
               <Separator className="w-5/12 bg-gray-600 mt-3 mx-2" /> or
               <Separator className="w-5/12 mt-3 pt-px mx-2 bg-gray-600 " />
