+++ tournament_app/components/lib/firebase/client-app.js
